feat: implement proper_of_time liturgical calendar generation
- Add ProperOfTime struct with comprehensive liturgical day generation
- Implement all liturgical seasons: Advent, Christmas Time, Lent, Paschal Triduum, Easter Time, Ordinary Time
- Add proper_of_time method to Preset struct for liturgical year calculation
- Add days command to CLI with YAML export and filtering capabilities
- Add LiturgicalDayFilter enum for property-based filtering
- Update schemas to include new liturgical day properties
- Add comprehensive test suite for calendar continuity and season-specific logic
- Fix week_of_season calculation across all seasons
- Add proper date ordering and duplicate handling
- Support epiphany_on_sunday configuration for flexible calendar variants

+++ core/src/liturgical_day.rs
@@ -83,7 +83,7 @@ pub struct LiturgicalDay {
 
     /// The nth occurrence of this day of the week within the current month.
     /// For example, the 3rd Sunday of the month would have nth_day_of_week_in_month = 3.
-    pub nth_day_of_week_in_month: u32,
+    pub nth_day_of_week_in_month: u8,
 
     /// The first day of the current liturgical season for this liturgical day.
     pub start_of_season: String, // in ISO 8601 format: YYYY-MM-DD
@@ -303,7 +303,7 @@ impl LiturgicalDay {
     }
 
     /// Sets the nth occurrence of this day of the week within the current month.
-    pub fn with_nth_day_of_week_in_month(mut self, nth: u32) -> Self {
+    pub fn with_nth_day_of_week_in_month(mut self, nth: u8) -> Self {
         self.nth_day_of_week_in_month = nth;
         self
     }
